Add links.js script to HTML files
