+++ about.html
@@ -111,5 +111,6 @@
         <p class="text-center text-muted">© 2023 KEA</p>
       </footer>
     </div>
+    <script src="js/links.js"></script>
   </body>
 </html>
