+++ admin.html
@@ -233,5 +233,6 @@
 
     <script src="js/main.js"></script>
     <script src="js/admin.js" type="module"></script>
+    <script src="js/links.js"></script>
   </body>
 </html>
