+++ event.html
@@ -121,5 +121,6 @@
     </div>
 
     <script src="js/event.js"></script>
+    <script src="js/links.js"></script>
   </body>
 </html>
