+++ events.html
@@ -82,5 +82,6 @@
     </div>
 
     <script src="js/events.js"></script>
+    <script src="js/links.js"></script>
   </body>
 </html>
