+++ index.html
@@ -91,5 +91,8 @@
         <p class="text-center text-muted">© 2023 KEA</p>
       </footer>
     </div>
+
+    <script src="js/links.js"></script>
+    <script src="js/links.js"></script>
   </body>
 </html>
