+++ js/auth.js
@@ -7,18 +7,5 @@ function checkUserToken() {
     }
 }
 
-export function getUserByToken(){
-    const localstorage_user = JSON.parse(localStorage.getItem('user'))
-    const options = {
-        method: 'GET',
-        headers: {
-            'Content-Type': 'application/json',
-            Authorization: `Bearer ${localstorage_user.token}`
-        }
-    }
-    return fetch(`${API}?token=${localstorage_user.token}`, options)
-    .then(res => res.json())
-}
-
 
 checkUserToken();
