+++ js/security.js
@@ -1,5 +1,3 @@
-import {getUserByToken} from "./auth.js";
-
 const API = 'http://localhost:8080/api/user';
 document.addEventListener("DOMContentLoaded", function () {
     document.getElementById('loginForm').addEventListener('submit', login);
@@ -41,6 +39,19 @@ function login(event) {
         });
 }
 
+
+function getUserByToken(){
+    const localstorage_user = JSON.parse(localStorage.getItem('user'))
+    const options = {
+        method: 'GET',
+        headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${localstorage_user.token}`
+        }
+    }
+    return fetch(`${API}?token=${localstorage_user.token}`, options)
+    .then(res => res.json())
+}
 
 function getToken(){
     const localstorage_user = JSON.parse(localStorage.getItem('user'))
