+++ login.html
@@ -99,6 +99,7 @@
       </footer>
     </div>
 
-    <script src="js/security.js"></script>
+    <script src="js/security.js" type="module"></script>
+    <script src="js/links.js"></script>
   </body>
 </html>
